Safari-611.1.10/LayoutTests/webaudio/audionode-expected.txt

Basic tests for AudioNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Source AudioNode has no inputs.
PASS Source AudioNode has one output.
PASS Destination AudioNode has one input.
PASS Destination AudioNode has no outputs.
PASS audioNode.connect(0, 0, 0) threw exception TypeError: Argument 1 ('destination') to AudioNode.connect must be an instance of AudioNode.
PASS audioNode.connect(null, 0, 0) threw exception TypeError: Argument 1 ('destination') to AudioNode.connect must be an instance of AudioNode.
PASS connect() exception thrown for illegal output index.
PASS connect() exception thrown for illegal input index.
PASS audioNode.connect(context.destination) succeeded.
PASS exception thrown when connecting to other context's node.
PASS exception should not be thrown when creating audio context with too many arguments.
PASS AudioNode is an EventTarget
PASS successfullyParsed is true

TEST COMPLETE
